+++ apps/researchers/peers/svc-rest-api/iac/gcp-cloud-run.tf
@@ -1,7 +1,9 @@
 # This resource block defines a Google Cloud Run service. This service will host the Docker image created by the Google Cloud Build trigger.
 resource "google_cloud_run_service" "apps_researchers_peers_rest_api" {
   # Name of the service
-  name = "${var.app_name}-${var.app_component_name}"
+  # name = "${var.app_name}-${var.app_component_name}" # Use the commit hash to force a new revision to be created
+  name = "${var.app_name}-${var.app_component_name}-${var.commit_hash}" # Use the commit hash to force a new revision to be created
+  # name = "${var.app_name}-${var.commit_hash}" # Use the commit hash to force a new revision to be created
 
   # The region where the service will be located
   location = var.region
@@ -15,7 +17,9 @@ resource "google_cloud_run_service" "apps_researchers_peers_rest_api" {
       # The Docker image to use for the service
       containers {
         # The docker image is pulled from GCR using the project ID, app name and the image tag which corresponds to the commit hash
-        image = "${var.region}-docker.pkg.dev/${var.project_id}/${var.gcp_docker_artifact_repository_name}/${var.app_name}:${var.commit_hash}"
+        # image = "${var.region}-docker.pkg.dev/${var.project_id}/${var.gcp_docker_artifact_repository_name}/${var.app_name}:${var.commit_hash}" # Use the commit hash to force a new revision to be created (this option forces a new docker build every time, which should not be necessary)
+        # image = "${var.region}-docker.pkg.dev/${var.project_id}/${var.gcp_docker_artifact_repository_name}/${var.app_name}:${var.environment}" # Use the environment to tag the image (staging, production, etc)
+        image = "${var.region}-docker.pkg.dev/${var.project_id}/${var.gcp_docker_artifact_repository_name}/${var.app_name}:latest" # Use the environment to tag the image (staging, production, etc)
 
         # Set the ENTRYPOINT_PATH environment variable (check the Dockerfile for more details)
         env {
@@ -58,10 +62,10 @@ resource "google_cloud_run_service" "apps_researchers_peers_rest_api" {
   }
 }
 
-# This block defines a Cloud Run IAM member. This sets the permissions for who can access the Cloud Run service.
-resource "google_cloud_run_service_iam_member" "public" {
-  service  = google_cloud_run_service.apps_researchers_peers_rest_api.name     # The name of the service to which the IAM policy will be applied
-  location = google_cloud_run_service.apps_researchers_peers_rest_api.location # The location of the service to which the IAM policy will be applied
-  role     = "roles/run.invoker"                                               # The role to be granted
-  member   = "allUsers"                                                        # The user, group, or service account who will have the role granted. In this case, all users.
-}
+# # This block defines a Cloud Run IAM member. This sets the permissions for who can access the Cloud Run service.
+# resource "google_cloud_run_service_iam_member" "public" {
+#   service  = google_cloud_run_service.apps_researchers_peers_rest_api.name     # The name of the service to which the IAM policy will be applied
+#   location = google_cloud_run_service.apps_researchers_peers_rest_api.location # The location of the service to which the IAM policy will be applied
+#   role     = "roles/run.invoker"                                               # The role to be granted
+#   member   = "allUsers"                                                        # The user, group, or service account who will have the role granted. In this case, all users.
+# }